Fix: Implementa regras de negócio no CarService

+++ DEVinCar.Service/Services/CarService.cs
@@ -24,7 +24,7 @@ namespace DEVinCar.Service.Services
                 query = query.Where(c => c.Name.ToUpper().Contains(name.ToUpper()));
 
             if (priceMin > priceMax)
-                throw new Exception(); //BadRequest()
+                throw new Exception(); // Minimal price cant be higher than maximum price
 
             if (priceMin.HasValue)
                 query = query.Where(c => c.SuggestedPrice >= priceMin);
@@ -33,36 +33,78 @@ namespace DEVinCar.Service.Services
                 query = query.Where(c => c.SuggestedPrice <= priceMax);
 
             if (!query.ToList().Any())
-                throw new Exception(); //NoContent()
+                throw new Exception(); // No car found
 
             return query.ToList();
         }
+
         public CarDTO GetById(int id)
         {
             return new CarDTO(_carRepository.GetById(id));
         }
-        // TODO
-        // Verificar se já existe um carro com esse nome return badRequest
-        // Verificar se o SuggestedPrice <= 0 return badRequest
+
         public void Post(CarDTO car)
         {
+            if (HasCarWithThisName(car.Name))
+                throw new Exception(); // Car with this name already exists
+
+            if (car.SuggestedPrice <= 0)
+                throw new Exception(); // Invalid car price
+
             _carRepository.Post(new Car(car));
         }
-        // TODO
-        // Verificar se carro existe se não retornar NotFound
-        // Verificar se carDto.Name.Equals(null) || carDto.SuggestedPrice.Equals(null) return BadRequest
-        // Verificar se carDto.SuggestedPrice <= 0 return BadRequest
-        // Verificar se nome do carro a editar já existe: Cars.Any(c => c.Name == carDto.Name && c.Id != carId) return BadRequest
+
         public void Alter(CarDTO car)
         {
+            if (CarNotFound(car.Id))
+                throw new Exception(); // Car not found
+
+            if (AllFieldsEmpty(car))
+                throw new Exception(); // Please fill all fields
+
+            if (HasDifferentCarWithThisName(car.Name, car.Id))
+                throw new Exception(); // Car with this name already exists 
+
+            if (car.SuggestedPrice <= 0)
+                throw new Exception(); // Invalid car price
+
+
             _carRepository.Alter(new Car(car));
         }
-        // TODO
-        // Verificar se o carro já está vendido retornar badRequest
+
         public void Delete(int carId)
         {
+            if (CarNotFound(carId))
+                throw new Exception(); // Car not found
+
+            if (_carRepository.HasBeenSold(carId))
+                throw new Exception(); // Cannot delete sold car
+
             Car car = _carRepository.GetById(carId);
             _carRepository.Delete(car);
+        }
+
+        private bool CarNotFound(int id)
+        {
+            return _carRepository.GetById(id) == null;
+        }
+
+        private bool HasCarWithThisName(string name)
+        {
+            return _carRepository.Get().Any(c => c.Name == name);
+        }
+
+        private bool HasDifferentCarWithThisName(string name, int carId)
+        {
+            return _carRepository.Get().Any(c => c.Name == name && c.Id != carId);
+        }
+
+        private bool AllFieldsEmpty(CarDTO car)
+        {
+            return (
+                String.IsNullOrEmpty(car.Name) &&
+                car.SuggestedPrice.Equals(null)
+            );
         }
     }
 }
